game/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.2.0
numpy>=1.19.5
scikit-learn>=0.24.2
matplotlib>=3.3.4
tqdm>=4.60.0